+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Remove the dimension input argument on GlobalPooling, Broadcast functions
 - Kernel map generation has tensor stride > 0 check
 - Fix `SparseTensor.set_tensor_stride`
+- Track whether the batch indices are set first when initializing coords, The initial batch indices will be used throughout the lifetime of a sparse tensor
 
 
 ## [0.4.1] - 2020-01-28